Use LegacyMessenger also in Drupal 8.
